requirements.txt
snowflake-connector-python
snowflake-snowpark-python
streamlit
pandas
